General defects correction (#21)
* refactor: changed integer to long

* refactor: removed unnecessary imports

+++ src/main/java/com/example/rebrain/controllers/TestController.java
@@ -1,11 +1,9 @@
 package com.example.rebrain.controllers;
 
-import com.example.rebrain.dto.CardDto;
 import com.example.rebrain.dto.TestAnswersDto;
 import com.example.rebrain.dto.TestDto;
 import com.example.rebrain.dto.TestPostDto;
 import com.example.rebrain.entity.TestEntity;
-import com.example.rebrain.mapper.CardMapper;
 import com.example.rebrain.mapper.TestMapper;
 import com.example.rebrain.services.TestService;
 import lombok.extern.slf4j.Slf4j;
@@ -40,7 +38,7 @@ public class TestController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<TestDto> getTestById(@PathVariable Integer id) {
+    public ResponseEntity<TestDto> getTestById(@PathVariable Long id) {
         log.debug("Fetching test with ID: {}", id);
         TestDto test = TestMapper.toDto(testService.getOne(id));
         log.debug("Test found: {}", test);
@@ -48,7 +46,7 @@ public class TestController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTest(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteTest(@PathVariable Long id) {
         log.info("Deleting test with ID: {}", id);
         testService.delete(id);
         log.info("Test with ID: {} deleted", id);
