+++ src/main/java/com/example/rebrain/dto/CardDto.java
@@ -1,6 +1,5 @@
 package com.example.rebrain.dto;
 
-import com.example.rebrain.entity.CardEntity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,5 +11,5 @@ public class CardDto {
     private Long id;
     private String title;
     private String description;
-    private Integer userId;
+    private Long userId;
 }
