+++ src/main/java/com/example/rebrain/dto/CardsSetsGetDto.java
@@ -1,6 +1,5 @@
 package com.example.rebrain.dto;
 
-import com.example.rebrain.entity.CardEntity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,7 +10,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CardsSetsGetDto {
-    private Integer setId;
+    private Long setId;
     private String setTitle;
     private String setDescription;
     private List<CardDto> cards;
