+++ src/main/java/com/example/rebrain/dto/SetGetDto.java
@@ -3,14 +3,13 @@ package com.example.rebrain.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class SetGetDto {
-    private Integer setId;
+    private Long setId;
     private String setTitle;
     private String setDescription;
     private List<CardDto> cards;
