+++ src/main/java/com/example/rebrain/dto/TestDto.java
@@ -3,17 +3,15 @@ package com.example.rebrain.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 import java.time.LocalDateTime;
-import java.util.Map;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class TestDto {
-    private Integer id;
-    private Integer setId;
-    private Integer userId;
+    private Long id;
+    private Long setId;
+    private Long userId;
     private String answers;
     private Integer cardsNumber;
     private Integer wrongAnswers;
