+++ src/main/java/com/example/rebrain/entity/Answer.java
@@ -1,6 +1,5 @@
 package com.example.rebrain.entity;
 
-import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Answer {
-    private Integer cardId;
+    private Long cardId;
     private String answer;
 }
