+++ src/main/java/com/example/rebrain/entity/CardsSetsId.java
@@ -1,16 +1,15 @@
 package com.example.rebrain.entity;
 
 import java.io.Serializable;
-import java.util.Objects;
 
 public class CardsSetsId implements Serializable {
-    private Integer cardId;
-    private Integer setId;
+    private Long cardId;
+    private Long setId;
 
     // Конструкторы
     public CardsSetsId() {}
 
-    public CardsSetsId(Integer cardId, Integer setId) {
+    public CardsSetsId(Long cardId, Long setId) {
         this.cardId = cardId;
         this.setId = setId;
     }
