+++ src/main/java/com/example/rebrain/repositories/NoteRepo.java
@@ -1,11 +1,9 @@
 package com.example.rebrain.repositories;
 
-import com.example.rebrain.entity.CardEntity;
 import com.example.rebrain.entity.NoteEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 
-public interface NoteRepo extends JpaRepository<NoteEntity, Integer> {
+public interface NoteRepo extends JpaRepository<NoteEntity, Long> {
     List<NoteEntity> findByTitle(String title);
 }
