+++ src/main/java/com/example/rebrain/repositories/SetRepo.java
@@ -1,11 +1,8 @@
 package com.example.rebrain.repositories;
 
-import com.example.rebrain.entity.CardEntity;
 import com.example.rebrain.entity.SetEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-public interface SetRepo extends JpaRepository<SetEntity, Integer> {
+public interface SetRepo extends JpaRepository<SetEntity, Long> {
 
 }
