+++ src/main/java/com/example/rebrain/repositories/UserRepo.java
@@ -3,7 +3,5 @@ package com.example.rebrain.repositories;
 import com.example.rebrain.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-public interface UserRepo extends JpaRepository<UserEntity, Integer> {
+public interface UserRepo extends JpaRepository<UserEntity, Long> {
 }
